// src/types/Notification.ts
export enum ENotification {
    none = "",
    waiting = "waiting",
    confirmed = "confirmed",
    noWallet = "noWallet",
    insufficientFunds = "insufficientFunds",
    inappropriateAmount = "inappropriateAmount",
    exceedingWithdrawal = "exceedingWithdrawal",
    insufficientAllowance = "insufficientAllowance",
}